Ver.1.0.0a26 :sparkles: log4jsを導入

+++ src/event/release-touch-data.ts
@@ -17,45 +17,31 @@ type ResponseType = void;
  * @param updateForce
  */
 export async function releaseTouchData(driver: Driver, exclusionOwner: string, arg: RequestType, updateForce?: boolean): Promise<ResponseType> {
-  console.log(`START [releaseTouchData (${exclusionOwner}) collection=${arg.collection} id=${arg.id}`);
+  const docSnap = await getData(driver, arg.collection, arg.id, { exclusionOwner });
 
-  let docSnap;
-  try {
-    docSnap = await getData(driver, arg.collection, arg.id, {
-      exclusionOwner
-    });
-  } catch (err) {
-    console.log(`ERROR [releaseTouchData (${exclusionOwner}) collection=${arg.collection} id=${arg.id}`);
-    throw err;
-  }
+  const createThrowDetail = (detail: string) => updateForce ? `Failure releaseTouchData. (${detail})` : detail;
 
-  if (!docSnap) {
-    console.log(`ERROR [releaseTouchData (${exclusionOwner}) collection=${arg.collection} id=${arg.id}`);
-    throw new ApplicationError(`Already released touch or created data. collection=${arg.collection} id=${arg.id} exclusionOwner=${exclusionOwner}`);
-  }
+  if (!docSnap) throw new ApplicationError(createThrowDetail("Already released touch or created."), arg);
 
-  try {
-    await deleteTouchier(driver, exclusionOwner, arg.collection, docSnap.ref.id);
-  } catch (err) {
-    console.log(`ERROR [releaseTouchData (${exclusionOwner}) collection=${arg.collection} id=${arg.id}`);
-    throw err;
-  }
+  await deleteTouchier(driver, exclusionOwner, arg.collection, docSnap.ref.id);
 
-  try {
-    if (updateForce || docSnap.data!.data) {
-      await docSnap.ref.update({
-        exclusionOwner: null,
-        updateTime: new Date()
-      });
-    } else {
-      await docSnap.ref.delete();
+  if (updateForce || docSnap.data!.data) {
+    const updateInfo = {
+      exclusionOwner: null,
+      updateTime: new Date()
+    };
+    try {
+      await docSnap.ref.update(updateInfo);
+    } catch (err) {
+      throw new ApplicationError(createThrowDetail("Failure update doc."), updateInfo);
     }
-  } catch (err) {
-    console.log(`ERROR [releaseTouchData (${exclusionOwner}) collection=${arg.collection} id=${arg.id}`);
-    throw err;
+  } else {
+    try {
+      await docSnap.ref.delete();
+    } catch (err) {
+      throw new ApplicationError(createThrowDetail("Failure delete doc."), arg);
+    }
   }
-
-  console.log(`END [releaseTouchData (${exclusionOwner}) collection=${arg.collection} id=${arg.id}`);
 }
 
 const resist: Resister = (driver: Driver, socket: any): void => {
